New PageRank implementation.
The page rank computation has been improved because we observed that
with the previous implementation there was really no conversion at all.

Now alfa and the number of iterations are taken from the command line
arguments.

Next step will be to devide the code in multiple files: one for each
class.